move endpoint definition to own module.
